\docType{data}
\name{base-patterns}
\alias{base-patterns}
\alias{complex.constant}
\alias{exp.constant}
\alias{following.characters}
\alias{int.constant}
\alias{name.pattern}
\alias{numeric.all.constant}
\alias{numeric.constant}
\alias{real.constant}
\alias{start.characters}
\title{Base Patterns}
\format{chr "[\\p{L}\\.]"}
\usage{
  start.characters

  following.characters

  name.pattern

  real.constant

  exp.constant

  int.constant

  complex.constant

  numeric.all.constant

  numeric.constant
}
\description{
  Use these perl regular expressions to help build pattern
  based styles.
}
\keyword{datasets}

