\name{rr}
\alias{.rr}
\alias{rr}
\title{make a results row

A convenience utility for creating the results rows for autotest style test
info}
\usage{
  .rr(line1, col1, line2, col2)
}
\description{
  make a results row

  A convenience utility for creating the results rows for
  autotest style test info
}
\keyword{internal}

