\name{test_pattern_style}
\alias{test_pattern_style}
\title{Test a patten for consistency.}
\usage{
  test_pattern_style(check, ti)
}
\arguments{
  \item{check}{a style check}

  \item{ti}{the test info data}
}
\value{
  either NULL or throws an error for use with test_that
}
\description{
  Test a patten for consistency.
}

