\name{with_default}
\alias{with_default}
\title{Look for an argument.}
\usage{
  with_default(x, default)
}
\arguments{
  \item{x}{an object}

  \item{default}{the default value}
}
\value{
  If x is neither NULL nor NA then x otherwise the default
}
\description{
  Look for an argument.
}

