\docType{data}
\name{assignment-styles}
\alias{assignment-styles}
\alias{styles.assignment.noeq}
\alias{styles.assignment.norightassign}
\alias{styles.assignment.notinfcall}
\title{Assignment Style}
\format{style tests for assignment operators.}
\usage{
  styles.assignment.noeq

  styles.assignment.norightassign

  styles.assignment.notinfcall
}
\description{
  Assignment Style
}
\keyword{datasets}

