\name{lint}
\alias{lint}
\title{Check a source document for stylistic errors.}
\usage{
  lint(file, text = NULL, tests = lint.tests)
}
\arguments{
  \item{file}{a vector of file paths.}

  \item{text}{text to check}

  \item{tests}{The list of tests to use to check.}
}
\description{
  Check a source document for stylistic errors.
}
\seealso{
  Other lint: \code{\link{check_pattern}}
}

