\docType{data}
\name{performance-styles}
\alias{performance.square}
\alias{performance-styles}
\title{Performace Enhancing Styles}
\format{Lint style checks.}
\usage{
  performance.square
}
\description{
  This collection of styles assert checks on known
  performance improvements.
}
\seealso{
  lint
}
\keyword{datasets}

